+++ Project.toml
@@ -1,0 +1,8 @@
+name = "Lens"
+uuid = "daab1da2-a660-11e8-1509-99065d28830a"
+authors = ["Zenna Tavares <zennatavares@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+Cassette = "7057c7e9-c182-5462-911a-8362d720325c"
+Spec = "526a04b8-654b-11e8-1588-db2a414f95b5"
