--- src/std.jl
@@ -1,5 +1,0 @@
-## Standard Listeners
-## ================
-
-printfl = Listener(:print, x->print(x...),true,false)
-printlnfl = Listener(:print, x->println(x...),true,false)
